Update mqtt.markdown
This page mentions "Home Assistant contains an embedded MQTT broker" but this contradicts the comment on this page https://home-assistant.io/getting-started/installation-raspberry-pi-all-in-one/ which mentions "Build and install Mosquitto v1.4.9 from source with websocket support running on ports 1883 and 9001" so it is very confusing for a NOOB as to what MQTT broker HASS is using. When I connect to 1883 and query the broker with mqtt.fx is reports it as Mosquitto 1.4.9.